+++ freud/cluster.pyx
@@ -27,34 +27,31 @@ cimport numpy as np
 np.import_array()
 
 cdef class Cluster(PairCompute):
-    R"""Finds clusters in a set of points.
+    """Finds clusters using a network of neighbors.
 
-    Given a set of coordinates and a cutoff, :class:`freud.cluster.Cluster`
-    will determine all of the clusters of points that are made up of points
-    that are closer than the cutoff. Clusters are 0-indexed. The class contains
-    an index array, the :code:`cluster_idx` attribute, which can be used to
-    identify which cluster a particle is associated with:
-    :code:`cluster_obj.cluster_idx[i]` is the cluster index in which particle
-    :code:`i` is found. By the definition of a cluster, points that are not
-    within the cutoff of another point end up in their own 1-particle cluster.
+    Given a set of particles and their neighbors,
+    :class:`freud.cluster.Cluster` will determine all of the connected
+    components of the network formed by those neighbor bonds. That is, two
+    points are in the same cluster if and only if a path exists between them on
+    the network of bonds. The class attribute :code:`cluster_idx` holds an
+    array of cluster indices for each particle. By the definition of a cluster,
+    points that are not bonded to any other point end up in their own
+    1-particle cluster.
 
-    Identifying micelles is one primary use-case for finding clusters. This
-    operation is somewhat different, though. In a cluster of points, each and
-    every point belongs to one and only one cluster. However, because a string
-    of points belongs to a polymer, that single polymer may be present in more
-    than one cluster. To handle this situation, an optional layer is presented
-    on top of the :code:`cluster_idx` array. Given a key value per particle
-    (i.e. the polymer id), the computeClusterMembership function will process
-    :code:`cluster_idx` with the key values in mind and provide a list of keys
-    that are present in each cluster.
+    Identifying micelles is one use-case for finding clusters. This operation
+    is somewhat different, though. In a cluster of points, each and every point
+    belongs to one and only one cluster. However, because a string of points
+    belongs to a polymer, that single polymer may be present in more than one
+    cluster. To handle this situation, an optional layer is presented on top of
+    the :code:`cluster_idx` array. Given a key value per particle (e.g. the
+    polymer id), the compute function will process clusters with the key values
+    in mind and provide a list of keys that are present in each cluster in the
+    attribute :code:`cluster_keys`, as a list of lists. If keys are not
+    provided, every particle is assigned a key corresponding to its index, and
+    :code:`cluster_keys` contains the particle ids present in each cluster.
 
     .. moduleauthor:: Joshua Anderson <joaander@umich.edu>
-
-    Args:
-        box (:class:`freud.box.Box`):
-            The simulation box.
-        r_max (float):
-            Particle distance cutoff.
+    .. moduleauthor:: Bradley Dice <bdice@bradleydice.com>
 
     .. note::
         **2D:** :class:`freud.cluster.Cluster` properly handles 2D boxes.
@@ -75,14 +72,14 @@ cdef class Cluster(PairCompute):
     cdef float r_max
 
     def __cinit__(self, float r_max):
-        self.thisptr = new freud._cluster.Cluster(r_max)
+        self.thisptr = new freud._cluster.Cluster()
         self.r_max = r_max
 
     def __dealloc__(self):
         del self.thisptr
 
-    @Compute._compute("compute")
-    def compute(self, box, points, nlist=None, query_args=None):
+    @Compute._compute()
+    def compute(self, box, points, keys=None, nlist=None, query_args=None):
         R"""Compute the clusters for the given set of points.
 
         Args:
@@ -90,6 +87,8 @@ cdef class Cluster(PairCompute):
                 Simulation box.
             points ((:math:`N_{particles}`, 3) :class:`np.ndarray`):
                 Particle coordinates.
+            keys ((:math:`N_{particles}`) :class:`numpy.ndarray`):
+                Membership keys, one for each particle.
             nlist (:class:`freud.locality.NeighborList`, optional):
                 Object to use to find bonds (Default value = :code:`None`).
         """
@@ -105,50 +104,39 @@ cdef class Cluster(PairCompute):
             self.preprocess_arguments(box, points, nlist=nlist,
                                       query_args=query_args)
 
+        cdef unsigned int* l_keys_ptr = NULL
+        cdef unsigned int[::1] l_keys
+        if keys is not None:
+            l_keys = freud.common.convert_array(
+                keys, shape=(num_query_points, ), dtype=np.uint32)
+            l_keys_ptr = &l_keys[0]
+
         self.thisptr.compute(
             nq.get_ptr(),
             nlistptr.get_ptr(),
-            <vec3[float]*> &l_query_points[0, 0],
-            num_query_points, dereference(qargs.thisptr))
-
-        return self
-
-    @Compute._compute("computeClusterMembership")
-    def computeClusterMembership(self, keys):
-        R"""Compute the clusters with key membership.
-        Loops over all particles and adds them to a list of sets.
-        Each set contains all the keys that are part of that cluster.
-        Get the computed list with :attr:`~cluster_keys`.
-
-        Args:
-            keys ((:math:`N_{particles}`) :class:`numpy.ndarray`):
-                Membership keys, one for each particle.
-        """
-        keys = freud.common.convert_array(
-            keys, shape=(self.num_particles, ), dtype=np.uint32)
-        cdef const unsigned int[::1] l_keys = keys
-        self.thisptr.computeClusterMembership(<unsigned int*> &l_keys[0])
+            dereference(qargs.thisptr),
+            l_keys_ptr)
         return self
 
     @property
     def default_query_args(self):
         return dict(mode="ball", r_max=self.r_max)
 
-    @Compute._computed_property("compute")
+    @Compute._computed_property()
     def num_clusters(self):
         return self.thisptr.getNumClusters()
 
-    @Compute._computed_property("compute")
+    @Compute._computed_property()
     def num_particles(self):
         return self.thisptr.getNumParticles()
 
-    @Compute._computed_property("compute")
+    @Compute._computed_property()
     def cluster_idx(self):
         return freud.util.make_managed_numpy_array(
             &self.thisptr.getClusterIdx(),
             freud.util.arr_type_t.UNSIGNED_INT)
 
-    @Compute._computed_property("computeClusterMembership")
+    @Compute._computed_property()
     def cluster_keys(self):
         cluster_keys = self.thisptr.getClusterKeys()
         return cluster_keys
@@ -158,7 +146,7 @@ cdef class Cluster(PairCompute):
                 "r_max={r_max})").format(cls=type(self).__name__,
                                          r_max=self.r_max)
 
-    @Compute._computed_method("compute")
+    @Compute._computed_method()
     def plot(self, ax=None):
         """Plot cluster distribution.